@@ -1,6 +1,6 @@
 {
   "name": "electron-hid-toy",
-  "version": "0.3.1",
+  "version": "0.4.0",
   "description": "Play with Electron and node-hid",
   "main": "main.js",
   "author": "Tod Kurt <todbotdotcom@gmail.com> (https://github.com/todbot)",
@@ -15,7 +15,8 @@
     "reload": "live-reload app --port 35729",
     "rebuild": "electron-rebuild",
     "electron-rebuild": "electron-rebuild",
-    "postinstall": "electron-builder install-app-deps"
+    "postinstall": "electron-builder install-app-deps",
+    "clean": "rimraf node_modules dist"
   },
   "keywords": [
     "hid",
@@ -48,11 +49,11 @@
     "babel-cli": "^6.10.1",
     "babel-preset-react": "^6.3.13",
     "babel-register": "^6.9.0",
-    "bootstrap": "^3.3.5",
+    "bootstrap": "^4.5.2",
     "node-hid": "^1.3.0",
-    "react": "^0.14.8",
-    "react-bootstrap": "^0.28.5",
-    "react-dom": "^0.14.7"
+    "react": "^16.13.1",
+    "react-bootstrap": "^1.3.0",
+    "react-dom": "^16.13.1"
   },
   "devDependencies": {
     "cross-env": "^6.0.3",
